requirements.txt
praw==7.7.1
openai>=1.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
tqdm>=4.66.0
